requirements.txt
fastapi
uvicorn
streamlit
reflex
sqlalchemy
pymysql
python-dotenv
cryptography
flet
requests
pandas
pytz